Move Comments API unit tests to proper location
- Moved unit tests from /tests/comments.rs to /src/endpoints/comments.rs
- Unit tests in Rust should be in the same module as the code they test
- /tests directory is for integration tests only
- Added #[cfg(test)] mod tests {} block to comments.rs with #[allow(clippy::expect_used)]
- All 18 tests still pass and are now properly located
